17h17 7th Nov

+++ backend/controllers/sauce.js
@@ -58,10 +58,23 @@ exports.modifySauce = (req, res, next) => {
 
 // Controller of the route PUT (delete one sauce)
 exports.deleteSauce = (req, res, next) => {
-    Sauce.deleteOne({ _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Sauce has been deleted !'}))
-        .catch(error => res.status(400).json({ error }));
-}
+    Sauce.findOne({ _id: req.params.id })
+        .then(sauce => {
+            if(sauce.userId !== req.auth.userId) {
+                res.status(401).json({ error: 'Unauthorized request !'});
+            } else {
+                const filename = sauce.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Sauce.deleteOne({ _id: req.params.id })
+                        .then(() => res.status(200).json({ message: 'Sauce has been deleted !'}))
+                        .catch(error => res.status(401).json({ error }));
+                });
+            }
+        })
+        .catch(error => {
+            res.status(500).json({ error });
+        });
+};
